% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compact-package.R
\docType{package}
\name{compact-package}
\alias{compact}
\alias{compact-package}
\title{compact: compact}
\description{
compact | Co-expression Module Perturbation Analysis in Cellular Transcriptomes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://smorabit.github.io/COMPACT/}
}

}
\author{
\strong{Maintainer}: Sam Morabito \email{smorabit@uci.edu} (\href{https://orcid.org/0000-0002-7768-4856}{ORCID})

Authors:
\itemize{
  \item Zechuan Shi \email{zechuas@uci.edu} (\href{https://orcid.org/0000-0002-2844-5816}{ORCID})
}

Other contributors:
\itemize{
  \item Swarup Lab [funder]
}

}
\keyword{internal}
