% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_plotting.R
\name{.create_distance_heatmap}
\alias{.create_distance_heatmap}
\title{This function generates a heatmap from a given matrix. It handles both
matrices and data frames (converting them to matrices if needed). It also allows
customization of the color palette, axis visibility, and legend display.}
\usage{
.create_distance_heatmap(
  df_matrix,
  title,
  min_val,
  max_val,
  custom_palette = NULL,
  show_x_axis = TRUE,
  show_y_axis = TRUE,
  show_legend = TRUE,
  custom_order = NULL
)
}
\arguments{
\item{df_matrix}{A numeric matrix or a data frame to be converted to a matrix.}

\item{title}{The title for the heatmap.}

\item{min_val}{The minimum value for the color scale.}

\item{max_val}{The maximum value for the color scale.}

\item{custom_palette}{A vector of colors to define the color palette. If \code{NULL}, a default palette is used.}

\item{show_x_axis}{Logical; whether to display the x-axis labels. Defaults to \code{TRUE}.}

\item{show_y_axis}{Logical; whether to display the y-axis labels. Defaults to \code{TRUE}.}

\item{show_legend}{Logical; whether to display the legend. Defaults to \code{TRUE}.}

\item{custom_order}{A character vector specifying the order of clusters for the x and y axes.}
}
\value{
A ggplot2 object representing the heatmap.
}
\description{
This function generates a heatmap from a given matrix. It handles both
matrices and data frames (converting them to matrices if needed). It also allows
customization of the color palette, axis visibility, and legend display.
}
\note{
This function is for internal use and is not exported.
}
