% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelExpression.R
\name{ModelZINB}
\alias{ModelZINB}
\title{ModelZINB}
\usage{
ModelZINB(
  seurat_obj,
  feature,
  cells_use = NULL,
  layer = "counts",
  slot = "counts",
  add_zero = TRUE
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{feature}{List of cell barcodes from the Seurat object}

\item{slot}{Slot to extract data for aggregation. Default = 'counts'}
}
\value{
The zero-inflated negative binomial model fit to the observed data from the selected feature.
}
\description{
This function models the expression of a selected feature as a zero-inflated negative binomial (ZINB) distribution.
}
\details{
ModelZINB is an internal helper function that calls on zeroinfl from the pscl package to
model expression of a selected feature.
}
