% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerturbationTransitions.R
\name{PerturbationTransitions}
\alias{PerturbationTransitions}
\title{PerturbationTransitions}
\usage{
PerturbationTransitions(
  seurat_obj,
  perturbation_name,
  features,
  graph,
  use_velocyto = TRUE,
  use_graph_tp = FALSE,
  corr_sigma = 0.05,
  n_threads = 4,
  layer = "data",
  slot = "data",
  assay = "RNA"
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{perturbation_name}{A name for the in-silico perturbation that will be stored in the Seurat obejct}

\item{features}{Selected features to use for the transition probability calculation}

\item{graph}{Name of the cell-cell graph in the Graphs(seurat_obj)}

\item{corr_sigma}{A numeric scaling factor for the correlation matrix.}

\item{n_threads}{Number of threads for the correlation calculation}

\item{slot}{Slot to extract data for aggregation. Default = 'data'}

\item{assay}{Assay in seurat_obj containing expression information.}
}
\value{
A Seurat object with the cell-cell transition probabilities stored in the Graphs slot of the Seurat object.
}
\description{
This function computes cell-cell transition probabilities based on an in-silico perturbation experiment.
}
