% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTransitionVectors.R
\name{PerturbationVectors}
\alias{PerturbationVectors}
\title{Calculate Perturbation Transition Vectors for Cell Embeddings}
\usage{
PerturbationVectors(
  seurat_obj,
  perturbation_name,
  reduction = "umap",
  n_threads = 4,
  arrow_scale = 1,
  max_pct = 0.9
)
}
\arguments{
\item{seurat_obj}{A Seurat object containing single-cell data, including embeddings and graphs.}

\item{perturbation_name}{A character string specifying the name of the perturbation. This is used to
retrieve the perturbation-specific transition probability graph.}

\item{reduction}{A character string specifying the name of the dimensional reduction to use for the
embedding (default: 'umap').}

\item{n_threads}{An integer specifying the number of threads to use for parallel processing
(default: 4).}

\item{arrow_scale}{A numeric value to scale the transition vectors, adjusting their length on the plot
(default: 1).}

\item{max_pct}{A numeric value between 0 and 1 indicating the maximum percentile of vector lengths
to display. This parameter caps the extreme values, limiting outlier vector lengths for better
visualization (default: 0.90).}
}
\value{
A list containing two data frames:
\item{ars}{A data frame with coordinates for plotting transition vectors, including initial points
(\code{x0}, \code{y0}) and final points (\code{x1}, \code{y1}) for each cell.}
\item{arsd}{A data frame with the transition vector distances (\code{xd}, \code{yd}) for each cell,
normalized based on \code{max_pct}.}
}
\description{
This function calculates transition vectors for cells in a Seurat object based on a specified
perturbation, providing insight into cell-state shifts on a dimensionality-reduced embedding.
}
\details{
This function retrieves the 2D embedding coordinates for the specified reduction in the Seurat
object. It then extracts the perturbation-specific transition probability graph from the Seurat object
and uses it to calculate transition vectors with the \code{embArrows_velocyto} helper function. Vector
lengths are scaled by \code{arrow_scale} and capped at the specified \code{max_pct} to handle extreme values.
Cells with \code{NA} values in either \code{ars} or \code{arsd} are excluded from the final output.
}
