% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictCellFates.R
\name{PredictCommitment}
\alias{PredictCommitment}
\title{Predict Fate Commitment (Committor Probabilities)}
\usage{
PredictCommitment(
  seurat_obj,
  perturbation_name,
  graph,
  source_cells = NULL,
  sink_cells = NULL,
  group.by = NULL,
  source_group = NULL,
  sink_group = NULL,
  output_name = "commitment_score",
  max_iter = 2000,
  tolerance = 1e-05,
  return_seurat = TRUE
)
}
\arguments{
\item{seurat_obj}{A Seurat object containing the required graphs.}

\item{perturbation_name}{Character. The base name of the simulated perturbation.
The function expects to find a transition probability graph named
\code{paste0(perturbation_name, '_tp')} in \code{seurat_obj@graphs}.}

\item{graph}{Character. The name of the K-Nearest Neighbors (KNN) graph stored in
\code{seurat_obj@graphs} (e.g., "RNA_nn") used to mask the transition probabilities.}

\item{source_cells}{Character vector. Cell barcodes to be defined as the starting
population (Boundary condition = 0).}

\item{sink_cells}{Character vector. Cell barcodes to be defined as the target
population (Boundary condition = 1).}

\item{group.by}{Character. The name of a column in \code{seurat_obj@meta.data}.
Used to automatically define source and sink cells if explicit lists are not provided.}

\item{source_group}{Character vector. The identity class(es) in \code{group.by}
representing the source state.}

\item{sink_group}{Character vector. The identity class(es) in \code{group.by}
representing the sink state.}

\item{output_name}{Character. The column name used to store the resulting probabilities
in \code{seurat_obj@meta.data}. Default is \code{"commitment_score"}.}

\item{max_iter}{Numeric. The maximum number of iterations for the solver. Default is \code{2000}.}

\item{tolerance}{Numeric. The convergence tolerance. Default is \code{1e-5}.}

\item{return_seurat}{Logical. If \code{TRUE}, returns the updated Seurat object.
If \code{FALSE}, returns a list containing the probabilities. Default is \code{TRUE}.}
}
\value{
If \code{return_seurat = TRUE}, a Seurat object with the calculated committor
probabilities appended to the metadata. If \code{return_seurat = FALSE}, a list
containing the probability scores.
}
\description{
Calculates the committor probability for every cell under a simulated perturbation.
This function models the vector field as a Dirichlet boundary value problem to answer
a specific question: "What is the exact probability that a cell will successfully
reach the target state (Sink) before it falls back into the origin state (Source)?"
This is highly effective for identifying epigenetic "points of no return" and
transcriptomic bottlenecks .
}
