% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelExpression.R
\name{SampleZINB}
\alias{SampleZINB}
\title{SampleZINB}
\usage{
SampleZINB(model, yobs, ncells = NULL)
}
\arguments{
\item{ncells}{number of cells to simulate expression values for. By default will simulate data for each cell in the input dataset.}
}
\value{
A numeric vector of simulated expression data based on a ZINB model
}
\description{
This function simulated expression values based on a ZINB model that has been fit to gene expression data.
}
