% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbation_functions.R
\name{ApplyPerturbation}
\alias{ApplyPerturbation}
\title{ApplyPerturbation}
\usage{
ApplyPerturbation(
  seurat_obj,
  exp,
  features,
  perturb_dir,
  cells_use = NULL,
  slot = "counts",
  assay = "RNA"
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{exp}{A features by cells matrix containing the observed expression matrix.}

\item{features}{The selected features to apply the perturbation on.}

\item{perturb_dir}{A numeric determining the type of perturbation to apply. Negative values for knock-down, positive for knock-in, and 0 for knock-out.}

\item{slot}{Slot to extract data for aggregation. Default = 'counts'}

\item{assay}{Assay in seurat_obj containing expression information.}
}
\value{
A dgCMatrix object containing the updated expression matrix with the applied perturbations
}
\description{
This function applies an in-silico perturbation to selected features in a Seurat object.
}
