% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_distance_test.R
\name{ComputeDistance}
\alias{ComputeDistance}
\title{ComputeDistance}
\usage{
ComputeDistance(
  seurat_object,
  groupby,
  reduction,
  method = "edist",
  verbose = TRUE
)
}
\arguments{
\item{seurat_object}{An object of class Seurat.}

\item{groupby}{An object of class character. Points to the column in the
Seurat object's meta data that contains the group labels.}

\item{reduction}{An object of class character. The reduction / embedding in
seurat_object that is used to compute the distance. Can be "pca", "harmony", or any linear dimensionality reduction.}

\item{method}{An object of class character. The distance calculation method, either "euclidean" or "edist".}

\item{verbose}{An object of class logical. If TRUE, prints messages.
Default is TRUE.}
}
\value{
Returns an object of class data.frame containing the distance between each pair of groups.
}
\description{
Computes either the Euclidean or E-distance between groups in a Seurat object, depending on the chosen method.
}
\examples{
compdist <- function(seurat_object, groupby = "Diagnosis", reduction = "pca", method = "euclidean", verbose = TRUE)
}
