% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_distance_test.R
\name{.eucldist}
\alias{.eucldist}
\title{eucldist}
\usage{
.eucldist(seurat_object, groupby, reduction, verbose = TRUE)
}
\arguments{
\item{seurat_object}{An object of class Seurat.}

\item{groupby}{An object of class character. Points to the column in the
Seurat object's meta data that contains the group labels.}

\item{reduction}{An object of class character. The reduction / embedding in
seurat_object that is used to compute the means. Can be "pca", "harmony", or any linear dimensionality reduction.}

\item{verbose}{An object of class logical. If TRUE, prints messages.
Default is TRUE.}
}
\value{
Returns an object of class data.frame containing the Euclidean distance between each pair of groups.
}
\description{
Computes the Euclidean distance between the means of two groups in a Seurat object.
}
\note{
This function is for internal use and is not exported. Computes the Euclidean distance between the means of two groups in a Seurat object.
}
